Move Merger and DuplicateCheck into database package
